+++ src/Properties/Changelog.md
@@ -7,6 +7,9 @@
 
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support
+- Changed the minimum Laravel 12 version to v12.61.1
+- Changed the minimum Laravel 13 version to v13.12
+- Added the `excerpt` and `description` fields to the `Property` model/table
 
 ## 5.x Series